Structure for creation of quests and npcs using only xml (no java code necessary). A few changes in chests classes were made due to undeleted debbug comments.
A few classes were implemented and they are basically the same for the
NPC and for the Quest. They are: structure class for storing
information, xml reader class and a generic class.

Two xml files were created too. One for npcs (speakerNPC.xml) and one
for quests (quests.xml).

Also, it was necessary to modify the way classes are loaded and add the
npc in semos city for testing.